+++ library/src/validations/notSize/notSize.ts
@@ -2,19 +2,19 @@ import type { ErrorMessage, PipeResult } from '../../types.ts';
 import { getOutput, getPipeIssues } from '../../utils/index.ts';
 
 /**
- * Creates a validation function that validates the length of a string or array.
+ * Creates a validation function that validates the size of a map, set or blob.
  *
- * @param requirement The maximum length.
+ * @param requirement The size.
  * @param error The error message.
  *
  * @returns A validation function.
  */
-export function maxLength<TInput extends string | any[]>(
+export function notSize<TInput extends Map<any, any> | Set<any> | Blob>(
   requirement: number,
   error?: ErrorMessage
 ) {
   return (input: TInput): PipeResult<TInput> =>
-    input.length > requirement
-      ? getPipeIssues('max_length', error || 'Invalid length', input)
+    input.size === requirement
+      ? getPipeIssues('not_size', error || 'Invalid size', input)
       : getOutput(input);
 }