+++ library/src/validations/notValue/notValue.test.ts
@@ -2,63 +2,27 @@ import { describe, expect, test } from 'vitest';
 import { notValue } from './notValue.ts';
 
 describe('notValue', () => {
-  test('should pass a string value that doesn’t match', () => {
-    const validate = notValue('foo');
+  test('should pass only valid values', () => {
+    const validate1 = notValue(3);
 
-    const result = validate('bar');
+    const value1 = 4;
+    expect(validate1(value1).output).toBe(value1);
+    const value2 = 2;
+    expect(validate1(value2).output).toBe(value2);
 
-    expect(result.output).toBe('bar');
-    expect(result.issues).toBeUndefined();
+    const validate2 = notValue('test');
+    const value3 = 'tes';
+    expect(validate2(value3).output).toBe(value3);
+    const value4 = 'testx';
+    expect(validate2(value4).output).toBe(value4);
+
+    expect(validate1(3).issues).toBeTruthy();
+    expect(validate2('test').issues).toBeTruthy();
   });
 
-  test('should pass a number value that doesn’t match', () => {
-    const validate = notValue(42);
-
-    const result = validate(21);
-
-    expect(result.output).toBe(21);
-    expect(result.issues).toBeUndefined();
-  });
-
-  test('should pass a bigint value that doesn’t match', () => {
-    const validate = notValue(42n);
-
-    const result = validate(21n);
-
-    expect(result.output).toBe(21n);
-    expect(result.issues).toBeUndefined();
-  });
-
-  test('should pass a boolean value that doesn’t match', () => {
-    const validate = notValue(true);
-
-    const result = validate(false);
-
-    expect(result.output).toBe(false);
-    expect(result.issues).toBeUndefined();
-  });
-
-  test('should not pass when value matches', () => {
-    const validate = notValue('foo');
-
-    const result = validate('foo');
-
-    expect(result.output).toBeUndefined();
-    expect(result.issues).toEqual([
-      {
-        input: 'foo',
-        message: 'Invalid value',
-        validation: 'not_value',
-      },
-    ]);
-  });
-
-  test('should return a custom error message when it doesn‘t pass', () => {
-    const error = 'the custom message';
-    const validate = notValue('foo', error);
-
-    const result = validate('foo');
-
-    expect(result.issues?.[0].message).toBe(error);
+  test('should return custom error message', () => {
+    const error = 'Value is "10"!';
+    const validate = notValue(10, error);
+    expect(validate(10).issues?.[0].message).toBe(error);
   });
 });