+++ library/src/validations/notValue/notValue.ts
@@ -2,7 +2,7 @@ import type { ErrorMessage, PipeResult } from '../../types.ts';
 import { getOutput, getPipeIssues } from '../../utils/index.ts';
 
 /**
- * Creates a validation functions that validates a value to be different from the given value.
+ * Creates a validation function that validates the value of a string or number.
  *
  * @param requirement The value.
  * @param error The error message.
@@ -10,11 +10,11 @@ import { getOutput, getPipeIssues } from '../../utils/index.ts';
  * @returns A validation function.
  */
 export function notValue<
-  TInput extends string | number | bigint | boolean,
+  TInput extends string | number | bigint,
   TRequirement extends TInput
->(requirement: TRequirement, error: ErrorMessage = 'Invalid value') {
+>(requirement: TRequirement, error?: ErrorMessage) {
   return (input: TInput): PipeResult<TInput> =>
     input === requirement
-      ? getPipeIssues('not_value', error, input)
+      ? getPipeIssues('not_value', error || 'Invalid value', input)
       : getOutput(input);
 }