+++ library/src/validations/uuid/uuid.ts
@@ -2,7 +2,7 @@ import type { ErrorMessage, PipeResult } from '../../types.ts';
 import { getOutput, getPipeIssues } from '../../utils/index.ts';
 
 /**
- * Creates a validation functions that validates a UUID.
+ * Creates a validation function that validates a UUID.
  *
  * @param error The error message.
  *